sched: freq: Reevaluate throttle if frequency requested changes
This is a bug that Andres found. Basically, if we're already in a down
throttle, any frequency up requests are ignored till the throttle period
ends. In this approach, we break out of any throttle if the requested
frequency changes and we check for the new request. Incase we are still
to throttle, we will continue once we return to the loop.

Test: uibench janktests
Bug: 33164692
Change-Id: Ife0ebd004ef7253a567cec81843d6cd049cae2fe
Signed-off-by: Joel Fernandes <joelaf@google.com>
(cherry picked from commit 7085a0b2b7e34fafe9e59d38ffb093d957e26bdd)