EXPERIMENTAL! sched/walt: set walt_disabled flag according to usage
Up until now, we have ignored the walt_disabled flag when you set/unset
sysctl_sched_use_walt_cpu/task_util. If this flag is cleared, then WALT
does not calculate anything.

Add a mechanism to catch writes to the sysctls and set walt_disabled
appropriately (and hence disable all the calculations) if neither are
in use. However, we do still wish to be able to perform comparisons between
WALT and PELT signals, so also add a scheduler feature to control this
behavior. If sched feature CALC_WALT is enabled (default) then we
always calculate WALT signals even if we are not using them.

This allows us to turn off the sched_feature and measure PELT as if
WALT were not present. However this has not been tested extensively so
beware of bugs if you turn it on.

Change-Id: I21b89ad4861857acef90eba6eda1cca661b34647
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
(cherry picked from commit 2da5e4c183cf088254bd2f7dcc7a7a30b71da920)